% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/app.R
\docType{package}
\name{csbCLUST-package}
\alias{csbCLUST}
\alias{csbCLUST-package}
\title{csbCLUST: Automated cluster detection from brightfield microscopy images}
\description{
The csbCLUST R package provides an interactive Shiny application
and a robust image-processing backend for automated detection,
quantification, and export of clusters from brightfield microscopy images.
}
\author{
\strong{Maintainer}: Caio Bonilha \email{your@email.com}

}
\keyword{internal}
